Refactor music_list and song_widget properly
